+++ packages/menu/src/MenuItem/MenuItem.styles.ts
@@ -7,6 +7,15 @@ export const disabledIconStyle: Record<Theme, string> = {
     color: ${palette.gray.dark2};
   `,
   [Theme.Dark]: css`
-    color: ${palette.gray.base};
+    color: ${palette.gray.dark1};
+  `,
+};
+
+export const destructiveIconStyle: Record<Theme, string> = {
+  [Theme.Light]: css`
+    color: ${palette.red.light1};
+  `,
+  [Theme.Dark]: css`
+    color: ${palette.red.light1};
   `,
 };