+++ packages/stepper/src/StepLabel/index.ts
@@ -1,0 +1,2 @@
+export { StepLabel } from './StepLabel';
+export { StepLabelProps } from './StepLabel.types';