+++ packages/marketing-modal/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @leafygreen-ui/marketing-modal
 
+## 4.1.2
+
+### Patch Changes
+
+- Updated dependencies [dd4f3da8]
+- Updated dependencies [90053e16]
+  - @leafygreen-ui/lib@13.0.0
+  - @leafygreen-ui/typography@18.0.0
+  - @leafygreen-ui/button@21.0.9
+  - @leafygreen-ui/leafygreen-provider@3.1.10
+  - @leafygreen-ui/modal@16.0.3
+
 ## 4.1.1
 
 ### Patch Changes